+++ StructuresAndAlgorithms/tree/tree/Classes/BinaryTree.cs
@@ -44,7 +44,7 @@ namespace tree.Classes
             {
                 PostOrderBottom(node.Right, values);
             };
-            values.Add(node.Value);
+            values.Add((int)node.Value);
 
             return values;
         }
@@ -73,7 +73,7 @@ namespace tree.Classes
             {
                 InOrderBottom(node.Left, values);
             }
-            values.Add(node.Value);
+            values.Add((int)node.Value);
             if (node.Right != null)
             {
                 InOrderBottom(node.Right, values);
@@ -101,7 +101,7 @@ namespace tree.Classes
         /// <returns> array of tree values in PreOrder </returns>
         private List<int> PreOrderBottom(Node node, List<int> values)
         {
-            values.Add(node.Value);
+            values.Add((int)node.Value);
             if (node.Left != null)
             {
                 PreOrderBottom(node.Left, values);
